refactor: merge refresh token
