+++ internal/handler/user_handler.go
@@ -231,3 +231,42 @@ func (h *UserHandler) Register(c *fiber.Ctx) error {
 	return c.Status(fiber.StatusCreated).JSON(res)
 
 }
+
+// Login godoc
+// @Summary Refresh user
+// @Description Refresh user
+// @Tags auth
+// @Accept json
+// @Produce json
+// @Param user body dto.RefreshTokenRequestBody true "user information"
+// @Success 200 {object} dto.SuccessResponse[dto.TokenResponseBody] "user successfully Refresh in"
+// @Failure 400 {object} dto.ErrorResponse "your request is invalid"
+// @Failure 401 {object} dto.ErrorResponse "your request is unauthorized"
+// @Failure 404 {object} dto.ErrorResponse "user not found"
+// @Failure 500 {object} dto.ErrorResponse "system cannot refresh user"
+// @Router /auth/refresh [post]
+func (h *UserHandler) RefreshToken(c *fiber.Ctx) error {
+	var req dto.RefreshTokenRequestBody
+	err := c.BodyParser(&req)
+	if err != nil {
+		return apperror.BadRequestError(err, "your request is invalid")
+	}
+	validate := validator.New()
+	if err := validate.Struct(req); err != nil {
+		return apperror.BadRequestError(err, "your request body is incorrect")
+	}
+
+	accessToken, refreshToken, loginErr := h.userService.RefreshToken(c.Context(), req.RefreshToken)
+	if loginErr != nil {
+		return loginErr
+	}
+
+	data := dto.TokenResponseBody{
+		AccessToken:  accessToken,
+		RefreshToken: refreshToken,
+	}
+
+	res := dto.Success(data)
+
+	return c.Status(fiber.StatusOK).JSON(res)
+}
